Refactoring: moved TestRunnerDecorator to parent package.
